+++ backend/user-service/src/main/java/com/example/userservice/entity/Location.java
@@ -21,11 +21,11 @@ public class Location extends TimeStamped {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private Float lat;
-    private Float lng;
+    private Double lat;
+    private Double lng;
     private Boolean moving;
 
-    public static Location of(User user, Float lat, Float lng, Boolean moving) {
+    public static Location of(User user, Double lat, Double lng, Boolean moving) {
         return Location.builder()
                 .user(user)
                 .lat(lat)
@@ -34,7 +34,7 @@ public class Location extends TimeStamped {
                 .build();
     }
 
-    public void updateLatLngMoving(Float lat, Float lng, Boolean moving) {
+    public void updateLatLngMoving(Double lat, Double lng, Boolean moving) {
         this.lat = lat;
         this.lng = lng;
         this.moving = moving;
